refactor: правки после ревью
