+++ src/common/constants.ts
@@ -1,0 +1,1 @@
+export const ITEMS_PER_PAGE = 3;
