+++ src/components/cart/MissingProdNotice.tsx
@@ -1,4 +1,4 @@
-import { Product } from "../../utils/types";
+import { Product } from "../../common/types";
 import { Button } from "react-bootstrap";
 import NoticeItem from "./NoticeItem";
 import List from "../List";
