--- src/utils/constants.ts
@@ -1,1 +1,0 @@
-export const itemsPerPage = 3;
